+++ packages/x-scheduler-internals-premium/src/internals/plugins/index.ts
@@ -1,2 +1,2 @@
 export * from './SchedulerLazyLoadingPlugin';
-export * from './SchedulerRecurringEventsPlugin';
+export * from './recurringEventsPlugin';
