+++ packages/x-scheduler-internals-premium/src/use-event-calendar-premium/EventCalendarPremiumStore.ts
@@ -1,7 +1,7 @@
 import { Adapter } from '@mui/x-scheduler-internals/use-adapter';
 import { ExtendableEventCalendarStore } from '@mui/x-scheduler-internals/use-event-calendar';
 import { EventCalendarPremiumLazyLoadingPlugin } from './plugins/EventCalendarPremiumLazyLoadingPlugin';
-import { SchedulerRecurringEventsPlugin } from '../internals/plugins/SchedulerRecurringEventsPlugin';
+import { recurringEventsPlugin } from '../internals/plugins/recurringEventsPlugin';
 import { EventCalendarPremiumParameters } from './EventCalendarPremiumStore.types';
 
 /**
@@ -13,16 +13,12 @@ export class EventCalendarPremiumStore<
 > extends ExtendableEventCalendarStore<TEvent, TResource> {
   public lazyLoading: EventCalendarPremiumLazyLoadingPlugin<TEvent>;
 
-  public recurringEventsPlugin: SchedulerRecurringEventsPlugin;
-
   public constructor(
     parameters: EventCalendarPremiumParameters<TEvent, TResource>,
     adapter: Adapter,
   ) {
-    const recurringEventsPlugin = new SchedulerRecurringEventsPlugin();
     super(parameters, adapter, 'EventCalendarPremiumStore', recurringEventsPlugin);
 
-    this.recurringEventsPlugin = recurringEventsPlugin;
     this.lazyLoading = new EventCalendarPremiumLazyLoadingPlugin<TEvent>(this);
   }
