+++ packages/x-scheduler-internals-premium/src/use-event-timeline-premium/EventTimelinePremiumStore.ts
@@ -14,7 +14,7 @@ import {
   EventTimelinePremiumParameters,
 } from './EventTimelinePremiumStore.types';
 import { EventTimelinePremiumLazyLoadingPlugin } from './plugins/EventTimelinePremiumLazyLoadingPlugin';
-import { SchedulerRecurringEventsPlugin } from '../internals/plugins/SchedulerRecurringEventsPlugin';
+import { recurringEventsPlugin } from '../internals/plugins/recurringEventsPlugin';
 import {
   EVENT_TIMELINE_PREMIUM_PRESET_CONFIGS,
   getPresetPxPerDay,
@@ -104,16 +104,11 @@ export class EventTimelinePremiumStore<
 > {
   public lazyLoading: EventTimelinePremiumLazyLoadingPlugin<TEvent>;
 
-  public recurringEventsPlugin: SchedulerRecurringEventsPlugin;
-
   public constructor(
     parameters: EventTimelinePremiumParameters<TEvent, TResource>,
     adapter: Adapter,
   ) {
-    const recurringEventsPlugin = new SchedulerRecurringEventsPlugin();
     super(parameters, adapter, 'EventTimelinePremiumStore', mapper, recurringEventsPlugin);
-
-    this.recurringEventsPlugin = recurringEventsPlugin;
 
     if (process.env.NODE_ENV !== 'production') {
       // Assert the initial state validity; `subscribe` only fires on subsequent state changes.
